Create SockMerchant_NoOfPairs.java
John works at a clothing store. He has a large pile of socks that he must pair by color for sale. Given an array of integers representing the color of each sock, determine how many pairs of socks with matching colors there are.

For example, there are 9 socks with colors 10 20 20 10 10 30 50 10 20. The number of pairs is 3 .